+++ src/main/java/sa/com/cloudsolutions/antikythera/evaluator/MockingEvaluator.java
@@ -14,12 +14,16 @@ import com.github.javaparser.ast.expr.NameExpr;
 import com.github.javaparser.ast.expr.NullLiteralExpr;
 import com.github.javaparser.ast.expr.ObjectCreationExpr;
 import com.github.javaparser.ast.expr.StringLiteralExpr;
+import com.github.javaparser.ast.stmt.ReturnStmt;
+import com.github.javaparser.ast.stmt.Statement;
 import com.github.javaparser.ast.type.ClassOrInterfaceType;
 import com.github.javaparser.ast.type.Type;
 import sa.com.cloudsolutions.antikythera.generator.TestGenerator;
 import sa.com.cloudsolutions.antikythera.parser.AbstractCompiler;
+import sa.com.cloudsolutions.antikythera.parser.Callable;
 
 import java.lang.reflect.Method;
+import java.util.List;
 import java.util.Optional;
 
 public class MockingEvaluator extends ControlFlowEvaluator {
@@ -67,10 +71,9 @@ public class MockingEvaluator extends ControlFlowEvaluator {
     }
 
     /**
-     *
      * @param cd CallableDeclaration a method being executed.
      * @return null when the method defined by the callable declaration is of the type void.
-     *   returns a reasonable value when the method is not void.
+     * returns a reasonable value when the method is not void.
      * @throws ReflectiveOperationException if the operation fails due to reflective issues
      */
     @Override
@@ -111,8 +114,7 @@ public class MockingEvaluator extends ControlFlowEvaluator {
             if (returnType.isClassOrInterfaceType() && returnType.asClassOrInterfaceType().getTypeArguments().isPresent()) {
                 String fqdn = AbstractCompiler.findFullyQualifiedName(cu1, returnType.asClassOrInterfaceType().getNameAsString());
                 result = Reflect.variableFactory(fqdn);
-            }
-            else {
+            } else {
                 String fqdn = AbstractCompiler.findFullyQualifiedName(cu1, returnType.toString());
                 result = Reflect.variableFactory(fqdn);
             }
@@ -161,65 +163,83 @@ public class MockingEvaluator extends ControlFlowEvaluator {
             case "List", "java.util.List", "java.util.ArrayList" -> {
                 TestGenerator.addDependency("java.util.ArrayList");
                 yield new ObjectCreationExpr()
-                    .setType(new ClassOrInterfaceType().setName("ArrayList"))
-                    .setArguments(new NodeList<>());
+                        .setType(new ClassOrInterfaceType().setName("ArrayList"))
+                        .setArguments(new NodeList<>());
             }
 
             case "Map", "java.util.Map", "java.util.HashMap" -> {
                 TestGenerator.addDependency("java.util.HashMap");
                 yield new ObjectCreationExpr()
-                    .setType(new ClassOrInterfaceType().setName("HashMap"))
-                    .setArguments(new NodeList<>());
+                        .setType(new ClassOrInterfaceType().setName("HashMap"))
+                        .setArguments(new NodeList<>());
             }
 
             case "java.util.TreeMap" -> {
                 TestGenerator.addDependency("java.util.TreeMap");
                 yield new ObjectCreationExpr()
-                    .setType(new ClassOrInterfaceType().setName("TreeMap"))
-                    .setArguments(new NodeList<>());
+                        .setType(new ClassOrInterfaceType().setName("TreeMap"))
+                        .setArguments(new NodeList<>());
             }
 
             case "Set", "java.util.Set", "java.util.HashSet" -> {
                 TestGenerator.addDependency("java.util.HashSet");
                 yield new ObjectCreationExpr()
-                    .setType(new ClassOrInterfaceType().setName("HashSet"))
-                    .setArguments(new NodeList<>());
+                        .setType(new ClassOrInterfaceType().setName("HashSet"))
+                        .setArguments(new NodeList<>());
             }
 
             case "java.util.TreeSet" -> {
                 TestGenerator.addDependency("java.util.TreeSet");
                 yield new ObjectCreationExpr()
-                    .setType(new ClassOrInterfaceType().setName("TreeSet"))
-                    .setArguments(new NodeList<>());
+                        .setType(new ClassOrInterfaceType().setName("TreeSet"))
+                        .setArguments(new NodeList<>());
             }
 
             case "java.util.Optional" -> {
                 TestGenerator.addDependency("java.util.Optional");
                 yield new MethodCallExpr(
-                    new NameExpr("Optional"),
-                    "empty"
+                        new NameExpr("Optional"),
+                        "empty"
                 );
             }
 
-            case "Boolean", Reflect.PRIMITIVE_BOOLEAN ->
-                new BooleanLiteralExpr(false);
+            case "Boolean", Reflect.PRIMITIVE_BOOLEAN -> new BooleanLiteralExpr(false);
 
             case Reflect.PRIMITIVE_FLOAT, Reflect.FLOAT, Reflect.PRIMITIVE_DOUBLE, Reflect.DOUBLE ->
-                new DoubleLiteralExpr("0.0");
+                    new DoubleLiteralExpr("0.0");
 
-            case Reflect.INTEGER, "int" ->
-                new IntegerLiteralExpr("0");
+            case Reflect.INTEGER, "int" -> new IntegerLiteralExpr("0");
 
-            case "Long", "long", "java.lang.Long" ->
-                new LongLiteralExpr("-100L");
+            case "Long", "long", "java.lang.Long" -> new LongLiteralExpr("-100L");
 
-            case "String", "java.lang.String" ->
-                new StringLiteralExpr("Ibuprofen");
+            case "String", "java.lang.String" -> new StringLiteralExpr("Ibuprofen");
 
-            default ->
-                new ObjectCreationExpr()
+            default -> new ObjectCreationExpr()
                     .setType(new ClassOrInterfaceType().setName(qualifiedName))
                     .setArguments(new NodeList<>());
         };
     }
+
+    @Override
+    Variable handleOptionals(ScopeChain.Scope scope) throws ReflectiveOperationException {
+        Callable callable = scope.getMCEWrapper().getMatchingCallable();
+        Method m = callable.getMethod();
+        Variable v = handleOptionalsHelper(scope);
+        if (v == null) {
+            return executeMethod(m);
+        }
+        return v;
+    }
+
+
+    @Override
+    Variable straightPath(ScopeChain.Scope sc, Statement stmt, MethodCallExpr methodCall) throws ReflectiveOperationException {
+        return null;
+    }
+
+    @Override
+    Variable riggedPath(ScopeChain.Scope sc, LineOfCode l) throws ReflectiveOperationException {
+        return null;
+    }
+
 }
